requirements.txt
python-telegram-bot[job-queue]==20.7
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
redis==5.0.1
pydantic-settings==2.1.0
